Added new infrarank type Forma (!)
Added support for a new infrarank type, or
Infratype, Forma, or "fma." for short.

Refactored all usage of the Infratype object
that directly use or make decisions based on
the subspecies or variety name no longer do
this, but instead more generically check the
infratype, and also pull data directly from
the infratype object.

Added a new column "code" to infratype table
that stores the three-letter code for each
infratype, this no longer needs to be decided
in code.

Finally, updated the library conversion script
to pull in the new forma infrarank.

   ALTER TABLE infratype ADD COLUMN code character varying(255);
   UPDATE infratype SET code = 'ssp.' WHERE id = 1;
   UPDATE infratype SET code = 'var.' WHERE id = 2;
   INSERT INTO infratype (id, name, code) VALUES (3, 'forma', 'fma.');

Ready to Deploy #567

Change-Id: I5aea67df655ce916daed7b89a1e761cdc18e267e
Reviewed-on: http://review.source.gogoego.com/2994
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>