Specify failure mode for MultiAppGenericBootstrap
Since some application may not be able to be installed
when starting multiple application, added a settable
flag to allow the extending class to specify how this
situation should be handled.

By default, a RuntimeException is thrown, but setting
this flag to false will allow the process to continue,
ignoring the failure.

Change-Id: I96d2b0b0aca6ab93e9e0c88fd00cc01617e8af96
Reviewed-on: http://review.source.gogoego.com/3006
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>