Working Sets load standalone, loading screens
WorkingSetCache now loads independently before other
cache data; this is needed because the multi-document
restlet simply fires off to asynchronous listeners,
which can not guarantee the working sets will be loaded
in time for when the navigation want to display them.
Other caches can continue to load asynchronously.

Also, added a few intermediate loading screen to show
when the working sets are loading versus cache data, and
also to display a welcome message when all data loading
is complete.

Finally, made the document processing use IncrementalCommand
to ensure that no long-running script warning will appear,
and also to promote update-able loading screens in the
future.

Change-Id: I9f7453e44456ae1d63599a26013b6d6ceae90934
Reviewed-on: http://review.source.gogoego.com/3010
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>