Re-mapped RedListCriteria data points (!).
The data point "dateLastSeen" has been re-named
to "yearLastSeen", and is now stored as a string
field instead of a date field.  While this data
point should store a year or range of years, the
UI will not attempt to constrict the user's input.

Also, while making improvements to this field,
changed the criteria version number to align
with the data in RedListCriteria_critVersionLookup.
The current version, 3.1 is now ID 1 instead of
0.  Each version ID has bumped up by one; joins
should now work on this table, and IDs start at
one, consistent with every other lookup table.
UI updates have been made to ensure that the
correct value is always saved, and pointers to
the IDs have been added to the field model (and
not hard-coded), though for necessary reasons, the
drop-down listing is still hard-coded.

Finally, changed the migration script to ensure
that criteria versions from SIS 1 get bumped up
by 1 (or set to 1 when null) to ensure that (a)
data for the current version lands in the database,
as has been deemed necessary, and (b) the values
map correctly to SIS 2.

Before deploying this change, the following queries
must be run; otherwise, there will appear to be
data loss for the dateLastSeen field, and all the
criteria versions will be displayed incorrectly:

   -- Updates to Lookups

   UPDATE "REDLISTCRITERIA"
   SET "NAME" = 'yearLastSeen', "DATA_TYPE" = 'string_primitive_field'
   WHERE "ID" = 10;

   -- Updates to databases

   SELECT 'INSERT INTO string_primitive_field (id, value) VALUES ('''|| f.id || ''',''' ||
	  substring(cast(f.value as character varying), 0, 5)
	  ||''');'
	FROM field
	JOIN primitive_field pf ON field.id = pf.fieldid
	  AND field.name = 'RedListCriteria' AND pf.name = 'dateLastSeen'
	JOIN date_primitive_field f ON pf.id = f.id

   -- Save the above query to file (no column names or quotes).
   -- Open and run the saved results as a query

   DELETE FROM date_primitive_field
   WHERE id in (
     SELECT f.id
     FROM field
     JOIN primitive_field pf ON field.id = pf.fieldid
       AND field.name = 'RedListCriteria' AND pf.name = 'dateLastSeen'
     JOIN string_primitive_field f ON pf.id = f.id
   );

   UPDATE primitive_field
   SET name = 'yearLastSeen'
   WHERE id in (
     SELECT f.id
     FROM field
     JOIN primitive_field pf ON field.id = pf.fieldid
       AND field.name = 'RedListCriteria' AND pf.name = 'dateLastSeen'
     JOIN string_primitive_field f ON pf.id = f.id
   );

   -- Remap criteria version to align with RedListCriteria_critVersionLookup

   UPDATE foreign_key_primitive_field
   SET value = 3
   WHERE id in (
     SELECT fk.id
     FROM field
     JOIN primitive_field pf ON field.id = pf.fieldid AND field.name = 'RedListCriteria' AND pf.name = 'critVersion'
     JOIN foreign_key_primitive_field fk ON pf.id = fk.id AND fk.value = 2
   );

   UPDATE foreign_key_primitive_field
   SET value = 2
   WHERE id in (
     SELECT fk.id
     FROM field
     JOIN primitive_field pf ON field.id = pf.fieldid AND field.name = 'RedListCriteria' AND pf.name = 'critVersion'
     JOIN foreign_key_primitive_field fk ON pf.id = fk.id AND fk.value = 1
   );

   UPDATE foreign_key_primitive_field
   SET value = 1
   WHERE id in (
    SELECT fk.id
    FROM field
    JOIN primitive_field pf ON field.id = pf.fieldid AND field.name = 'RedListCriteria' AND pf.name = 'critVersion'
    JOIN foreign_key_primitive_field fk ON pf.id = fk.id AND fk.value = 0
   );

Ready to Deploy #576

Change-Id: Ibc1098de470a2ab50078f3a7cd53c668dabfb2be
Reviewed-on: http://review.source.gogoego.com/3012
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>