Adjust summary display for CountryOccurrence
Widened the data field columns by 100 pixels to allow
for room for all options to display on one line, and
shortened the Country column a bit since, typically,
this is not very long.  Some will still go on two lines,
but not the majority.

Also, removed Formely Bred from the listing, since the
value is typically "Unknown".  This column can still
be viewed by clicking the "View/Edit" button.

Ready to Deploy #574

Change-Id: I012a2d46621026491ce0218e1593fc31cbbb4816
Reviewed-on: http://review.source.gogoego.com/3013
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>