Simple fields, narratives support size attribute.
Added support for a size attribute for simple fields,
such as numbers, text, dates, and most importantly,
narratives (text areas).

The narrative implementation of size allows for five
different keywords that affect the size of the textarea:

* small (100x50)
* *medium (200x100)*
* large (400x200)
* xl (500x300)
* xxl (600,400)

Setting the size attribute to any of these keywords for
narrative fields or justification fields will set the
text area to the denoted sizes (in pixels).  For these
field types, medium is the default size.

See #575

Change-Id: Ie8ac357c57104c68a10c5725166d4f9f827b0931
Reviewed-on: http://review.source.gogoego.com/3014
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>