Lazy-load definitions, initialize before usage.
Definitions are not a core piece of the SIS system
that need to be loaded and initialized on startup
like other caches.  Removed the HasCache interface
markup and instead, have definitions lazily loaded
on demand, either by using the definition editor
panel or requesting to see a definition when using
the Assessment Data Browser, slimming application
initialization time.

Additionally, made a code-cleanup sweep to remove
odd behaviors -- the window will no longer open if
there are no definiable terms found, it will simply
state this in a warning message.  Also, the names
will appear in the original case, instead of being
forced to lowercase, and they will also appear in
alphabetical order.

Change-Id: I3fa909948c69b6e2875e4a54a09745a76bc09a42
Reviewed-on: http://review.source.gogoego.com/3037
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>