TransactionRestlet doesn't start trx for GET reqs
For GET requests, a transaction is no longer started.
By nature, requests that are GETting data should have
no need for having a data-transformating transaction
open.  The session is still opened, but no transaction
gets committed (or rolled back) upon method completion.

Should review this for side effects, but overall, if
any code breaks due to this commit, that code is
probably written wrong and should be corrected.

Change-Id: I9ba0be2e8cf2167d7014a4c333151831edb121f5
Reviewed-on: http://review.source.gogoego.com/3038
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>