WorkingSetCache respect FetchMode operations
FetchMode and typed CacheEntry objects from AssessmentCache
have been pulled out into the caching package and are now
utilized in WorkingSetCache.

The WorkingSetCache pulls PARTIAL data for working sets when
listing all of a user's working sets in the navigation, and
fetches the entire set when a user selects it in the nav or
it is pre-loaded via bookmark.  While certain operations do
check to ensure this data is fetched before-hand, any kind
of operation done on a working set can be safely assumed to
have a FULL working set loaded by the time the code reaches
that point.

Loading the PARTIAL working set with the listing operation
removes a large load time for parsing working sets, especially
those with a large number of taxa and subscribers, since this
information is not transferred from the server.  While the
PARTIAL listing could become even more lightweight, the changes
here are sufficient to make a noticeable dent in the load time.

Additionally, this commit fixes a couple cache-related bugs:

* Assessments now get removed from the navigation upon deletion
* StateManager accepts event arguments for the reset method
* Working Set subscriptions stick.
* Working Set -> Assessment cache cleared upon add/rm assessment

Ready to Deploy #572

Change-Id: Iec922adea8795308dc8371ddd3157f47fccf6b72
Reviewed-on: http://review.source.gogoego.com/3039
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>