Error throw/catch fix for locked assessments.
First, server was wrapping writing exceptions with proper
locking error messages as server exceptions.  So, fixed
this so ResourceExceptions are re-thrown as themselves, and
also ensure that SAVE_LOCK returns Forbidden instead of Locked
to be consistent with other locks types.

Next, client was taking 500 errors (which are possible if
an unhandled exception occurs) and doing nothing with them,
so added a block to handle these errors.  This was the original
problem, and while this commit prevents that from occuring again,
the actual issue (locking) should now be presented.

Ready to Deploy #583

Change-Id: I7bc813399c7506a06e868e07b75fd10a1018c3cf
Reviewed-on: http://review.source.gogoego.com/3083
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>