Moved all hand-written SQL to abstract queries
Created an abstraction layer for queries that must be
hand-written (typically ones relating to working sets).
Some of them may (or may not) have database-vendor-
specific SQL, thus SIS will now ensure that the right
implementation for SQL queries is being used for a given
vendor.

Note that only Postgresql is being handled here, we will
want to make one for H2 (though the queries have not
had issues so far).

Change-Id: Ie780277eacd805d5b23085822b146dd6d3ee0530
Reviewed-on: http://review.source.gogoego.com/3089
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>