Revamped PermissionGroup editor, code cleanup.
Revamped the permission group editor and surrounding
functionality to correct and add the following:

* Corrected grantable working set listing
* Corrected rename functionality
* Corrected sorting algorithm
* Simplified persistence layer / process
* Standardized UI bits with similar control types
* Added "Summarize" feature: PermissionGroup.toString()
* Swapped drop-down to checkboxes for easier reading

This, among other small tweaks and code cleanup.  Last
item of note here -- this panels comes out of a tab panel
and is instead in its own window; the menu changes to a
drop-down now.

The old editor is still in-tact and available in
debug mode for comparison purposes.

Ready to Deploy #573

Change-Id: I40c60503a7b08af0859cddfe42e2d3527a5b9cda
Reviewed-on: http://review.source.gogoego.com/3088
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>