Fixed save for permissions in user management.
Corrected an error that was causing only the visible
permission groups to persist -- some permission group
(related to working sets) would go away when saved
since they were not visible.

Also, made the username column sortable and sorted
the permission group listing alphabetically and put
the checked ones at the top of the list.

Ready to Deploy #587
Ready to Deploy #588
Ready to Deploy #589

Change-Id: I2596c59b134962dbbeed3b6e8f8f753ab32ee373
Reviewed-on: http://review.source.gogoego.com/3092
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>