Bug Fix: Auto-save causes widgets to lose focus
The standard refresh function for the navigation causes
the rows to grab focus.  This is not needed when doing
a refresh to simply update text, and is in fact bad for
the assessment data browser.

Changed the refresh function to take a flag that will let
the caller specify if the selection should be updated
(which is not needed on a simple refresh), and changed
the HeaderContainer refresh calls to use the simple
version.  Since the grid no longer grabs focus, for the
assessment data browser specifically, widgets will no
longer lose focus.

Ready to Deploy #592

Change-Id: Idcd3f16da2be2bccad92bd3c46302913923d4503
Reviewed-on: http://review.source.gogoego.com/3111
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>