Ensure FuzzyResults up-to-date, explicit re-run.
Added a means to explicitly re-run the expert system
via the Criteria Results window.  Also, when a run
is done, ensure that the new fields is properly added
to the assessment and accessible via the mapping cache.

Finally, removed RedListFuzzyResults as a field that
gets carried over when creating an assessment from
a template; hence, this field will be re-generated.

Ready to Deploy #596

Change-Id: Ic68074900a10273d9c9794ae6d76e34fcf4529db
Reviewed-on: http://review.source.gogoego.com/3113
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>