Taxon Basic Editor cleanup, corrections.
Cleaned up the interface for the taxon basic
editor, removed duplicate code, and removed
the save and close option -- save, by default,
now closes the window.

Also, corrected issues with changes not immediately
appearing in the client and friendly names not being
updated properly on the server after changes are
made.

Finally, added Forma infratype to the Level drop-down.

Ready to Deploy #598
Ready to Deploy #600

Change-Id: Ide0fc5193471decad597c6f30c5c0ff87aa68988
Reviewed-on: http://review.source.gogoego.com/3114
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>