Restore All option for WorkingSet perms, more...
Had got lost in the shuffle of a previous commit.
Always appearing at the top, this will set the
permission url to resource/workingSet.

Also, ordered the list of features in the drop-down
and display a friendly-name description instead of
the cryptic key.

Finally, removed some fixed FIXMEs from HeaderContainer.

Change-Id: Ifd59d607965dd7e9b788eb01470ede3a2ddfb95d
Reviewed-on: http://review.source.gogoego.com/3154
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>