Correct Working Set Assessments Sorting Order
The list of assessments for a particular working sets
in the client was only showing assessments for the
selected taxon when attempting to iterate through them
on the Assesment Data Browser page.

Changed this to not use the selected taxon, causing the
proper behavior.  However, ordering was an issue here, so
updated the server logic to sort the navigable taxa in the
correct order such that iterator is consistent with the
navigation.  In doing this, moved the applicable
comparators out to the bridge project; should follow suit
with other comparators going forward.

Ready to Deploy #607

Change-Id: I4feb42fa7085c3bceef0b6e2c74b838d72ec791d
Reviewed-on: http://review.source.gogoego.com/3157
Tested-by: Carl Scott <carl.scott@solertium.com>
Reviewed-by: Carl Scott <carl.scott@solertium.com>