SearchCache removes duplicate entries upon search.
When a search is performed, the SearchCache will now
cache by query; if found, if it's a newly-typed search,
the cached entry will be removed and a new search will
be performed to get fresh results, and the queried item
will move to the bottom of the list in the Gear menu.
Otherwise, as in the case of selecting a search from the
Gear menu, the search results will be displayed based on
the cached results.

Ready to Deploy #612

Change-Id: Iefe88d340ec8ab9a0771fc5e0ad5939295de2de0
Reviewed-on: http://review.source.gogoego.com/3166
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>