Taxon Assessment Listing Grid fix, cleanup
Fixed an error causing the grid to not appear when
a user did not have appropriate rights to read a
published assessment.

Also, cleaned up code a bit -- removed row listener
and hand-coded HTML images in favor of IconButtons,
and moved functionality to their own methods for
ease of reading and maintaining.

Ready to Deploy #617

Change-Id: I9dddf2c5114fba78a049647109695b0ef0b6ab7d
Reviewed-on: http://review.source.gogoego.com/3189
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>