StateManager now validates contents onChange.
When a call is made to a helper state-setting function,
the WorkingSet will now be interrogated to ensure that
the object changing state (either a taxon or assessment)
is within the scope of the selected working set, if any.
If not, the selected working set will be deselected, thus
correcting the URL.

Also, for basic search, searching by ID will now always
take you directly to the taxon page, even if the taxon
is in the selected working set.  This ensures consistent
behavior here.

Ready to Deploy #619

Change-Id: Id20331036c0bc47931c42603905105f4ecfecbb2
Reviewed-on: http://review.source.gogoego.com/3191
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>