Access Export Support, DB Scratch Template
Corrected issues with AccessExport and added a download
resource to get the zipped access database.  Updated the
working set page to show the content panel similar to
the export to offline, and prompt the user before running.

Added a scratch (empty) access database to use as a template.
This can be updated in the future with something more
interesting.

Ready to Deploy #85

Change-Id: Ica7779f039e338b10c7f4fc20eaa60bb02ac6f76
Reviewed-on: http://review.source.gogoego.com/3196
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>