Classification Scheme data sync fixes, delete fix.
Changed Classification Scheme save logic to completely
overwrite any field information with the data stored in
the actual models.  This should be safe since the data
is sync'd with the server for each save.  Corrections
were also made to that very sync process to ensure all
data is synchronized properly.

This should ensure no duplicate are sent to the server
to be re-saved, and also ensures that what's on-screen
is what truly gets saved.

Also, corrected an issue in Threats (caused by bad data)
where a Stress could become null, and throw an error.
This data will now not only not throw the error, but
should also clean itself up over time.

Additionally, for Threats, updated the models to
ensure that the lookup table names are stored with
the data.

Finally, fixed an issue where deleting single entries
were not marking the classification scheme viewer as
dirty, thus that change alone could not be saved.

Ready to Deploy #623

Change-Id: I50514519d630dc72cf93bb48189962aca1e6463e
Reviewed-on: http://review.source.gogoego.com/3199
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>