Assessment and Working Set Navigation cleanup
For assessment panel, lock icon will now appear for
working sets that a user does not have rights to
read.  Fixes a long-standing TODO.

For working sets, removed the download icon, as this
is working-set specific and can be done from the
working set page.  Also, cleaned up the code for the
information panel.

Change-Id: I5d4851c209e096814d3f8be7bacc6e8724c1d688
Reviewed-on: http://review.source.gogoego.com/3203
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>