Trx Restlet API update to allow for trx management
TransactionRestlet now supports a means of the implementing
restlet determing whether a session or a transaction is opened
when viewing data.

By default, all method calls open a transaction except when
the method is GET.  It's assumed a GET will not be performing
any changes...

This logic has also been ported to TransactionResource.

Change-Id: Ib5fa31c25b83356d64d34eb6d6bd744dcd22820c
Reviewed-on: http://review.source.gogoego.com/3224
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>