Display descriptions for SISTreeStructures.
Ensure that the description is shown on the SISTreeStructure.
This is probably not a long-term thing, as there should be a
way to suppress this, but is fine for now since all use-cases
for trees that would need to supress descriptions use a
ClassificationScheme, not a Field, to display.

However, this affects Use/Trade, since Classification Schemes
are embedded in the one to many fields; therefore, the labels
should show.

See #238

Change-Id: I219524f7067c11c3c272a20e0f720d731492494f
Reviewed-on: http://review.source.gogoego.com/3229
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>