Display upgrades for SISOneToMany structures.
Changed the one-to-many structure to use popup
widnows and listings to display records instead of
an infinitely-scrolling page.  These records shall
now also support notes and references, and should
be easier to view and navigate the data.

The components involved in saving and loading have
remained the same, this is more of a facelift,
with new functionality added.

While not complete, this commit should be considered
a "status quo" offering -- that is, things should
work just as well as they did before.  There are still
some tweaks left in regard to notes & references and
some naming conventions.

See #238

Change-Id: I7be5e0c84da7029388d3a58eeb0036bc4ed72947
Reviewed-on: http://review.source.gogoego.com/3230
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>