Corrected and reinstated primitive field deletion.
Reinstated support for primitive field deletion in the
AssessmentChangesRestlet after making a few corrections and
additions to the AssessmentPersistence layer.

The sink method for PrimitiveFields in AssessmentPersistence
now uses the field name as the key instead of the field id
when attempting to synchronize data.  Since the primitive
field name is contractually obligated to be unique, both in
practice and via the equals method, this should work.  Also,
it will prevent sending different versions of the same field
down.

Also, added support in AssessmentChangePacket for notes; these
notes are server-side only at this time, but can be added when
data is being merged in order to put more contextual information
into the change packet.  When the change packet is written to
file, these notes are included. Now, when a field or primitive
field is deleted (typically due to having no data), this deletion
is logged and the contents are displayed as well (in addition to
it being in the database-backed change log).  This allows quick
reference to see what was deleted and why.

Finally, fixed an issue with saving global references and changed
the save mode such that an edit is added to the assessment when
global references are edited.

Ready to Deploy #627

Change-Id: Ied386d6b92e4dbf7b74bd440bab86515f4f0f6c4
Reviewed-on: http://review.source.gogoego.com/3256
Tested-by: Carl Scott <carl.scott@solertium.com>
Reviewed-by: Carl Scott <carl.scott@solertium.com>