Offline client launcher, build file, cleanup.
Created Offline.launch to launch the client in offline
mode.  Updated Offline.gwt.xml to use a base configuration
shared with the online version since they use much of
the same config.  Also, added helper methods to the SIS
object for isOnline and isOffline, refactoring where the
old means of detection were used in order to make the code
more readable.

Added in a placeholder offline toolbar when running in
offline mode (which does nothing now, but is on the screen).

Finally, added a build file that will build the offline
version of the client.

Ready to Deploy #642
Fixed #643
See #644

Change-Id: I5ec44e5b2fbcf4668c306fa6fd590898fad8de1b
Reviewed-on: http://review.source.gogoego.com/3261
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>