Fix for hasChanged for multi-select fields.
The hasChanged method can trigger when multiple selections
for a field are ordered differently when written to string.
For example, selections of 1, 3, and 5 could be written as
"1,3,5" or "1,5,3" -- these are the same, but the hasChanged
logic thinks they are different.

Updated this logic so that the string always comes out in
order, both from the UI field and in the getRawValue()
method.

Change-Id: Ia1b225a02f5995c6bf674f82338f6692053c3d57
Reviewed-on: http://review.source.gogoego.com/3262
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>