Additional styling for OneToMany page links.
The page links now have a blue vertical bar alongside the
left of the widget to help make this more noticeable
amongst other fields.  Also, the containing panel's
components are now spaced 10px apart to give it a bit more
whitespace, helping it stand out a tad more and growing
the height of the vertical bar.

See #651

Change-Id: I2b1b415ed1089b729978e4685e9cc8350073f545
Reviewed-on: http://review.source.gogoego.com/3284
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>