Remove special-case code for Use/Trade model (!)
There were a number of places in the code that were written
especially to account for some nuances with the old model for
Use/Trade.  Now that those nuances are much better understood,
those special cases have been removed, opting instead to fix
the core of the problem first, and trickling up from there.

The UseTradeSubfield has been renamed to UseTradeDetailsSubfield,
in the first part of this effort, for consistency with other fields
that work like this (classification schemes, Livelihoods).  From
here, FIXME were removed from LookupLoading classes, as it should
just work.

This will also fix the viewmaster views for UseTrade, as they did
not adjust for the Use/Trade case; now, they don't need to.

Last on this front was correcting the client tools to ensure that
the right data is being saved to the appropriate fields with the
appropriate field names, using the field Model objects to help
make these assurances.

Also in this commit are small fixes to the creates & inserts SQL
for these tables and a couple others.

Finally, the following SQL will need to run in order to ensure
that the existing Use/Trade data does not "disappear" when this
commit is deployed:

   UPDATE field SET name = 'UseTradeDetailsSubfield' WHERE name = 'UseTradeSubfield';

The lookups table should also be re-generated based on the latest
creates & inserts SQL files.

Ready to Deploy #667

Change-Id: I9c498ca4c78671d3b31007b3b84cb9d9ebdc6306
Reviewed-on: http://review.source.gogoego.com/3343
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>