Publication Target CRUD, UI Management.
UI Management panel for Publication Targets, allowing users
to add new targets from the manager and on-the-fly when using
batch update for publication data.  Add/edit require a name
and date, with a publication reference being optional.  For
delete, the operation will only work if there is no data tied
to the target.

Also, in this commit, the publication restlet has been broken
into two separate restlets, one for data, one for targets, and
the PublicationTarget model now lazily holds the PublicationData
it's targeting.  This is useful for queries and also pre-checking
delete operations.

Publication Target management is, for now, available under the
administrative tools in the Gear menu.

Ready to Deploy #675

Change-Id: I7837b890b17326a3404b2de9352d62c2a19fb871
Reviewed-on: http://review.source.gogoego.com/3404
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>