Recent Activities Service and Portlet
Beta version of a recent activities portlet, which displays
the most recent 250 activites for that day.  Activities are
currently limited to assessment changes, but I envision this
working for any type of edit.

Also, clicking a title currently takes you to the assessment
that was changed, but future implementation could have this
load the change itself, with the diff.  Also, formatting is
a bit rough around the edges, but it shows the correct
content.

See #276

Change-Id: I7fa7f44b65cdaeac5bb3d56f939f0da9747709d9
Reviewed-on: http://review.source.gogoego.com/3406
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>