Administrative tools now load in modal window.
All administrative tools now load in a modal popup
window.  Also, added a means of setting the size of
this window specifically for windows that need a
bit more room.

To this note, set the size of the publication management
screen to be 800x600 (100px larger on both ends).

Ready to Deploy #684

Change-Id: Ifb6799950b89ec9fcc4f31fedffd344e0f68c2f7
Reviewed-on: http://review.source.gogoego.com/3411
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>