Rollback Changes from Track Changes Panel.
Changes can now be rolled back on a field-by-field
basis from the track changes panel.  For this initial
version, classification schemes are not yet available
for rollback.

Additionally, an API update now allows a save to choose
a merging method after the save; Merge does the normal
sync of the IDs, whereas Replace bodily replaces the
assessment object in the cache.  Upon rollback, a
save/replace is done, followed by a forced refresh of
the assessment data browser.

Change-Id: Ia9debba50b1f386622e37f2aa324ec89be264eff
Reviewed-on: http://review.source.gogoego.com/3415
Tested-by: Carl Scott <carl.scott@solertium.com>
Reviewed-by: Carl Scott <carl.scott@solertium.com>