Implemented Publication via Publication Manager
When setting the status to "Published" in the publication
workflow manager, upon save, data publication will occur
on each selected assessment -- that being the items listed
in #464.  Code for this was moved from AssessmentIO to
PublicationIO, and the TODO regarding PublicationTarget and
references was removed, since this is now available.

Also, corrected a listing issue with the data list filtering
out draft assessments instead of published ones, and corrected
a view updating issue when saving RLU notes.

See #685
Ready to Deploy #686

Change-Id: I67ffd24dbdd1b1cc7ca790f9a013e1b3631ba0c0
Reviewed-on: http://review.source.gogoego.com/3417
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>