Bug Fix: Reference column "year" broke Hibernate
On certain setups and configurations, the column "year"
in the reference table was causing issues as, apparently,
it was a reserved word on that platform.  This is also
the case for the user table, but a fix was already applied.

The fix for this, for future reference, is documented here:

   https://hibernate.onjira.com/browse/HHH-5135

"If a column has the same name as a no-arg function, (e.g.,
current_date), then the mapped column will need to have
back-ticks (e.g., `current_date`) to work properly."

Hence, this commit updated Reference.hbm.xml to set the year
column declaration as "`year`".

Again, this only happens on certain setups, apparently, and
has not been an issue since it was put in, so more checks
may be need to be done on the problem platform.

Local testing shows this does not break working systems.

Change-Id: I731bc61dec98ce9e58b5d033349b4bbdc8c0b1ea
Reviewed-on: http://review.source.gogoego.com/3424
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>