WorkingSet Taxon fetch optimization, request wait.
Optimized the request made to fetch taxa for working sets
such that taxa now come back fully with working sets when
a FULL fetch is made for a single working set. This means
that full taxa can be cached when a working set is loaded
to the client so a subsequent (and also long-running)
request to the server is not made.  This can save up to
four round-trips to the server.

The taxon listing will still attempt to make a request to
the server based on the information in its cache -- so if
the taxon is reached via URL jump or some other means, it
will still load appropriately, since the taxa for listing
in the navigation are not yet cached, thus a server fetch
will happen as it currently does.

Also, added a request-wait protocol to working set cache
such that any GET request made for a working set will now
first check to see if that request is already in progress;
if so, the callback will be put into a cache with a list
of waiting callbacks until the request in progress is
complete.  Once that happens, all waiting requests will
be callback back.  This helps ensure that a double-click
on the working set name in the navigation does not cause
two loads -- one for the navigation and another for the
working set page -- both of which retrieve the same data.

Combined, these two efforts in tandem should be noticable
improvements to the user experience since it take away 4
to 5 requests that used to happen per-click.

Ready to Deploy #691

Change-Id: I42c8594e0ef6813df807189fe2b7dbf4d68790c4
Reviewed-on: http://review.source.gogoego.com/3427
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>