Cleanup Taxomatic Tools Menu on Taxon Home Page
Removed unused options from the Taxomatic Tools Menu,
specifically the undo feature.  This may or may not
return in the future, so code is only commented out.

Also, apply permissions to both the Editing features
for a taxon (formerly "Taxon Tools") as well as the
overall Taxomatic features for the menu.  The menu
will appear differently based on permissions.

Finally, changed check for enabling taxomatic features
to specifically check for online/offline status. However,
this still does not remove the functionality from the
server, which should still happen in #646.

See #646
Ready to Deploy #694

Change-Id: Id9755ed19abfb891410b99684f1db156bbf45493
Reviewed-on: http://review.source.gogoego.com/3432
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>