Offline: Corrections for JSNI status, parsing.
Corrected an issue in the OfflineMetadataRestlet with parsing
the file name, as it was pulling from an unused property name.
Set it to the correct property name.

Also, changed the lastModified property of OfflineMetadata from
a string to a Date object, to allow for more manipulication of
this object on the client.

Finally, corrected the JSNI code responsible for triggering
updates when the browser is determine to be online/offline, and
set the label in the toolbar to update when these events occur.
Unfortunately, this code may have to change in the near future
due to shortcomings in certain browser implementations, but the
approach in use now is correct.

See #644

Change-Id: I3441e68d6c851935ed27fa1c3b1fb3d7704ba61a
Reviewed-on: http://review.source.gogoego.com/3436
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>