Recent Activity logic update for My Working Sets.
The option for listing activity on taxa within ones working
sets originally fell back to list all if users did not have
any working sets.  However, with the drop-down now available
in the client to choose an option, as well as the default of
these options being ALL, logic has been changed back to have
My Working Sets return an empty set if there are no working
sets for that user.

Also, corrected the join order for the working set listing.

Fixed #690

Change-Id: I0071545099a9e32ec7945991827428dbf35359b0
Reviewed-on: http://review.source.gogoego.com/3439
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>