Offline support RecentActivity via canned queries.
Added SQL from RecentActivityRestlet to the canned query API
in order to provide support for this functionality offline (H2);
the date comparison portion of the query is distinct from
Postgresql.

In doing so, added a distinct H2CannedQueries implementation and
applied this to be used when an H2 database connection is detected.

Ready to Deploy #700

Change-Id: I137f9086639840eb79a48d53fed994201db68ac8
Reviewed-on: http://review.source.gogoego.com/3444
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>