Changed prop name for SIS->Hibernate ID generator.
The property name to set the generator type for Hibernate
is now:

   org.iucn.sis.server.configuration.generator

as opposed to its former name of simply "generator."  Now
that it is no longer for internal use only, the name is
important.

This also relates to fixing the save issues with the offline
H2 database for SIS.  It seems that changing the generator
type from "native" to "increment" has allowed the IDs to
correctly generate in the H2 database, thus, data can now
be saved.

Test #699

Change-Id: Idf9a399b4f27aaef8d4144972215b1d450359621
Reviewed-on: http://review.source.gogoego.com/3446
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>