Change Hibernate CFG parsing protocol.
In SISPersistentManager, the HBM files were read in and
parsed as Documents, which was going very slowly on some
configurations.  Changed this to instead read in as a
string, while still performing the replacements that were
being done originally.  Result is a faster load time for
these files.

The old method is still available as an option to use, but
the string-based method is currently in use.

Change-Id: Ie7a5e41921cb9f9c26e81a1f573f94d0b89c8583
Reviewed-on: http://review.source.gogoego.com/3453
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>